Disable automatic update checks in gimp
Summary: The latest version of gimp added the default setting to automatically phone home and check for updates. We are handling updates ourselves and thus there are not much of a reason of having network requests every time the application starts.

Test Plan: - Check for updates inside the about dialog.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D8319